[REFACTOR]: id값 Int -> Long으로 수정

+++ app/src/main/java/com/kuit/ourmenu/data/model/map/response/MapDetailResponse.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class MapDetailResponse(
     @SerialName("menuId")
-    val menuId: Int,
+    val menuId: Long,
     @SerialName("menuTitle")
     val menuTitle: String,
     @SerialName("menuPrice")
